Removing RadioConfig trait
- As functions in the RadioConfig trait could take different types based
  on what you can configure, making it generic seemed too complex.
  We'll re-assess the right way to do this, but for now we're just going
  to remove it.
- Renamed rtlsdr_node to rtlsdr_radio since we have generic RadioNodes
- Updated tests and fixed clippy issues.
