+++ src/hardware/mod.rs
@@ -2,6 +2,6 @@
 extern crate rtlsdr;
 
 #[cfg(feature = "rtlsdr_support")]
-pub mod rtlsdr_node;
+pub mod rtlsdr_radio;
 
 pub mod radio;
